Fix bug

+++ src/deposit.h
@@ -1,6 +1,6 @@
 #ifndef DEPOSIT_H
 #define DEPOSIT_H
 
-int analiz (float d, int s);
+int analiz (int d, float s);
 
 #endif
